requirements.txt
streamlit==1.48.1
pandas==2.2.2
numpy==1.26.4
scikit-learn
scipy
scikit-surprise
